refactor: Dealer와 Player가 Participant를 상속받도록 수정
